Bugfix: Cell IDs wrongly assigned in interactive sessions
So far we have assumed that the a path would always be added to the most recently
added Tree in the Path Manager. That is a reasonable assumption when adding paths 
in bulk but not during an interactive session. When adding a branch, this fix checks for the tree ID  of the parent path and assigns it to the branched path