Remove "Make Primary" command
* "Set Active Node As Tree Root" in Edit Mode is the preferred command to re-root a Tree, as it allows exact choice of root.
* Remove Path#unsetPrimaryForConnected

This method no longer works as intended due to the removal of the primary flag. It is also no longer necessary since the "primary" property of a Path is now determined by its connectivity (i.e., the primary Path is the single Path in a connected set of Paths where startJoins == null). It may be worth re-introducing this flag if there is a computational benefit to not finding the set of primary Paths every time you need them.