Fix #147
Unless specified, String.format() uses the current locale to format doubles, which means that a comma
may be used as decimal separator. This ensures that a '.' is always used to ensure consistency when saving
file. There are still other String.format() calls that do not specify locale, but those seem to be used only in
debugging/log calls and thus, not as consequential.