Path: replace old array-based design with single collection
This is a major refactoring 10 years in the making: All of the arrays holding path
properties (coordinates, colors, values, etc.) are now handled by a single list.
Tests are passing but it is likely that something will brake.
This all provides a definite fix for the hacks in place when importing SWC (e.g., 2e14451bde).
Quick tests suggest that
handling large files 2k+ nodes is ~1.5-2x times faster!!